+++ app/gui/src/dashboard/layouts/InfoMenu.tsx
@@ -1,6 +1,7 @@
 /** @file A menu containing info about the app. */
 import { PRODUCT_NAME } from 'enso-common'
 
+import { LOGIN_PATH } from '#/appUtils'
 import LogoIcon from '#/assets/enso_logo.svg'
 import { Popover, Text } from '#/components/AriaComponents'
 import MenuEntry from '#/components/MenuEntry'
@@ -11,8 +12,7 @@ import { useAuth } from '#/providers/AuthProvider'
 import { useSetModal } from '#/providers/ModalProvider'
 import { useSessionAPI } from '#/providers/SessionProvider.tsx'
 import { useText } from '#/providers/TextProvider'
-import { useNavigate } from 'react-router-dom'
-import { LOGIN_PATH } from '../appUtils'
+import { useRouterInReact } from '$/providers/react'
 
 /** Props for an {@link InfoMenu}. */
 export interface InfoMenuProps {
@@ -23,7 +23,7 @@ export interface InfoMenuProps {
 export default function InfoMenu(props: InfoMenuProps) {
   const { hidden = false } = props
 
-  const navigate = useNavigate()
+  const { router } = useRouterInReact()
   const { signOut } = useSessionAPI()
   const { session } = useAuth()
   const { setModal } = useSetModal()
@@ -51,11 +51,7 @@ export default function InfoMenu(props: InfoMenuProps) {
             {session && (
               <MenuEntry
                 action="signOut"
-                doAction={() =>
-                  signOut().then(() => {
-                    navigate(LOGIN_PATH)
-                  })
-                }
+                doAction={() => signOut().then(() => router.push(LOGIN_PATH))}
               />
             )}
           </div>