+++ app/gui/src/dashboard/modals/InviteUsersModal/InviteUsersSuccess.tsx
@@ -1,14 +1,13 @@
 /** @file Success screen for the "invite users" modal. */
 import * as React from 'react'
 
-import * as reactRouterDom from 'react-router-dom'
-
 import ArrowRightIcon from '#/assets/arrow_right.svg'
 
 import * as textProvider from '#/providers/TextProvider'
 
 import * as ariaComponents from '#/components/AriaComponents'
 import * as result from '#/components/Result'
+import { useRouterInReact } from '$/providers/react'
 
 /**
  * The number of emails to display in the success message.
@@ -33,7 +32,7 @@ export function InviteUsersSuccess(props: InviteUsersSuccessProps) {
     ['cloud-ide_SettingsTab', '"members"'],
   ] as const
 
-  const [searchParams, setSearchParams] = reactRouterDom.useSearchParams()
+  const { route, router } = useRouterInReact()
 
   const emailListFormatter = React.useMemo(
     () => new Intl.ListFormat(locale, { type: 'conjunction', style: 'long' }),
@@ -41,8 +40,8 @@ export function InviteUsersSuccess(props: InviteUsersSuccessProps) {
   )
 
   const isUserOnMembersPage =
-    searchParams.has(membersSearchParams[0][0], membersSearchParams[0][1]) &&
-    searchParams.has(membersSearchParams[1][0], membersSearchParams[1][1])
+    route.query[membersSearchParams[0][0]] === membersSearchParams[0][1] &&
+    route.query[membersSearchParams[1][0]] === membersSearchParams[1][1]
 
   return (
     <result.Result
@@ -71,10 +70,8 @@ export function InviteUsersSuccess(props: InviteUsersSuccessProps) {
               onPressStart={onClose}
               onPress={() => {
                 onClose()
-                membersSearchParams.forEach(([key, value]) => {
-                  searchParams.set(key, value)
-                })
-                setSearchParams(searchParams)
+                const newQuery = { ...route.query, ...Object.fromEntries(membersSearchParams) }
+                void router.push({ query: newQuery })
               }}
             >
               {getText('goToMembersPage')}